Reorganizing Test Mocks structure
