+++ nasa_apod_app/lib/core/external/external.dart
@@ -2,4 +2,4 @@ export 'helpers/errors/external_error_type.dart';
 export 'helpers/errors/external_error_type_extension.dart';
 export 'helpers/errors/external_exception.dart';
 export 'services/clients/http_client_adapter.dart';
-export 'services/storages/local_storage_adapter.dart';
+export 'services/device/local_storage_adapter.dart';
