+++ nasa_apod_app/lib/core/external/services/device/local_storage_adapter.dart
