--- nasa_apod_app/test/mocks/core/external/device/adapters/adapters.dart
@@ -1,1 +1,0 @@
-export 'storages/local_storage_spy.dart';
