--- nasa_apod_app/test/mocks/core/external/device/device.dart
@@ -1,2 +1,0 @@
-export 'adapters/adapters.dart';
-export 'directories/device_local_storage_factory.dart';
