+++ nasa_apod_app/test/mocks/core/external/external.dart
@@ -1,3 +1,5 @@
-export 'device/device.dart';
-export 'objects/picture_factory.dart';
-export 'remote/remote.dart';
+export 'services/clients/api/apod_responses_factory.dart';
+export 'services/clients/http_client_spy.dart';
+export 'services/device/directories/device_local_storage_factory.dart';
+export 'services/device/local_storage_spy.dart';
+export 'services/objects/picture_factory.dart';
