--- nasa_apod_app/test/mocks/core/external/remote/adapters/adapters.dart
@@ -1,1 +1,0 @@
-export 'clients/http_client_spy.dart';
