--- nasa_apod_app/test/mocks/core/external/remote/remote.dart
@@ -1,2 +1,0 @@
-export 'adapters/adapters.dart';
-export 'api/apod_responses_factory.dart';
