+++ nasa_apod_app/test/mocks/core/external/services/clients/api/apod_responses_factory.dart
@@ -1,6 +1,6 @@
 import 'package:nasa_apod_app/nasa_apod_app.dart';
 
-import '../../../../../apod.dart';
+import '../../../../../../apod.dart';
 
 class ApodResponsesFactory with PictureFactory {
   InfraErrorType generateNotFoundError() => InfraErrorType.notFound;
