+++ nasa_apod_app/test/mocks/core/external/services/clients/http_client_spy.dart
