+++ nasa_apod_app/test/mocks/core/external/services/device/directories/device_local_storage_factory.dart
@@ -1,4 +1,4 @@
-import '../../../../../apod.dart';
+import '../../../../../../apod.dart';
 
 class DeviceLocalStorageFactory with PictureFactory {
   List<Map<String, dynamic>> generateIncompletePictureMapList() => [
