+++ nasa_apod_app/test/mocks/core/external/services/device/local_storage_spy.dart
