+++ nasa_apod_app/test/mocks/core/external/services/objects/picture_factory.dart
